Fixes #389 Created ignite/utils.py with public methods from ignite/_utils.py
